controllers: Move Sources::write to a separate controller
