+++ index.php
@@ -46,8 +46,8 @@ $f3->route('GET /sources', 'controllers\Sources->show'); // html
 $f3->route('GET /source', 'controllers\Sources->add'); // html
 $f3->route('GET /sources/list', 'controllers\Sources->listSources'); // json
 $f3->route('GET /sources/sourcesStats', 'controllers\Sources->sourcesStats'); // json
-$f3->route('POST /source/@id', 'controllers\Sources->write'); // json
-$f3->route('POST /source', 'controllers\Sources->write'); // json
+$f3->route('POST /source/@id', 'controllers\Sources\Write->write'); // json
+$f3->route('POST /source', 'controllers\Sources\Write->write'); // json
 $f3->route('DELETE /source/@id', 'controllers\Sources->remove'); // json
 $f3->route('POST /source/delete/@id', 'controllers\Sources->remove'); // json
 $f3->route('POST /source/@id/update', 'controllers\Sources\Update->update'); // json
