+++ src/controllers/Sources.php
@@ -130,119 +130,6 @@ class Sources {
     }
 
     /**
-     * render spouts params
-     * json
-     *
-     * @param Base $f3 fatfree base instance
-     * @param array $params query string parameters
-     *
-     * @return void
-     */
-    public function write(Base $f3, array $params) {
-        $f3->get('auth')->needsLoggedIn();
-
-        $sourcesDao = new \daos\Sources();
-
-        // read data
-        $headers = \F3::get('HEADERS');
-        $body = \F3::get('BODY');
-        if (isset($headers['Content-Type']) && strpos($headers['Content-Type'], 'application/json') === 0) {
-            $data = json_decode($body, true);
-        } else {
-            parse_str($body, $data);
-        }
-
-        $data['spout'] = str_replace('_', '\\', $data['spout']);
-        if (!isset($data['title']) || strlen(trim($data['title'])) === 0) {
-            // try to fetch title, if it is not filled in
-            $loader = new \helpers\ContentLoader();
-            $title = $loader->fetchTitle($data);
-
-            if ($title) {
-                $data['title'] = $title;
-            } else {
-                $this->view->jsonError(['title' => 'no title given and could not fetch it']);
-            }
-        }
-
-        if (!isset($data['spout'])) {
-            $this->view->jsonError(['spout' => 'no data for spout given']);
-        }
-
-        // clean up title and tag data to prevent XSS
-        $title = htmlspecialchars($data['title']);
-        $tags = array_map('htmlspecialchars', $data['tags']);
-        $spout = $data['spout'];
-        $filter = $data['filter'];
-
-        unset($data['title']);
-        unset($data['spout']);
-        unset($data['filter']);
-        unset($data['tags']);
-
-        // check if source already exists
-        $id = $params['id'];
-        $sourceExists = $sourcesDao->isValid('id', $id);
-
-        // load password value if not changed for spouts containing passwords
-        if ($sourceExists) {
-            $spoutLoader = new \helpers\SpoutLoader();
-            $spoutInstance = $spoutLoader->get($spout);
-
-            foreach ($spoutInstance->params as $spoutParamName => $spoutParam) {
-                if ($spoutParam['type'] === 'password'
-                    && empty($data[$spoutParamName])) {
-                    if (!isset($oldSource)) {
-                        $oldSource = $sourcesDao->get($id);
-                        $oldParams = json_decode(html_entity_decode($oldSource['params']), true);
-                    }
-                    $data[$spoutParamName] = $oldParams[$spoutParamName];
-                }
-            }
-        }
-
-        $validation = $sourcesDao->validate($title, $spout, $data);
-        if ($validation !== true) {
-            $this->view->error(json_encode($validation));
-        }
-
-        // add/edit source
-        if (!$sourceExists) {
-            $id = $sourcesDao->add($title, $tags, $filter, $spout, $data);
-        } else {
-            $sourcesDao->edit($id, $title, $tags, $filter, $spout, $data);
-        }
-
-        // autocolor tags
-        $tagsDao = new \daos\Tags();
-        foreach ($tags as $tag) {
-            $tagsDao->autocolorTag(trim($tag));
-        }
-
-        // cleanup tags
-        $tagsDao->cleanup($sourcesDao->getAllTags());
-
-        $return = [
-            'success' => true,
-            'id' => (int) $id,
-            'title' => $title
-        ];
-
-        // only for selfoss ui (update stats in navigation)
-        if ($f3->ajax()) {
-            // get new tag list with updated count values
-            $tagController = new \controllers\Tags($this->authentication, $this->view);
-            $return['tags'] = $tagController->tagsListAsString();
-
-            // get new sources list
-            $sourcesController = new \controllers\Sources($this->authentication, $this->view);
-            $return['sources'] = $sourcesController->sourcesListAsString();
-        }
-
-        $this->view->jsonSuccess($return);
-    }
-
-    /**
      * return source stats in HTML for nav update
      * json
      *
